+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # v3.0.0-rc.4
-## mm/dd/2019
+## 03/20/2019
 
 1. [](#improved)
     * Make fields `formname`, `uniqueid` and `honeypot` non-inputs in form validation
     * Update all Form classes to rely on `PageInterface` instead of `Page` class
     * Removed `media.upload_limit` references
+    * Added field type `hidden` to `ip` and `timestamp` actions
 1. [](#bugfix)
     * Fixed null date/time in list view